created a approve customer page and added its component to employee dashboard
